refactor(back): Refactor news interface.
I move the definition of the news type into `graphql/type/news.js` rather than
`graphql/data_types/dataTypes.js`. I also move the query field `getNews` and
`createNews` into `graphql/type/news.js`.

I remove `graphql/query/query.js` because it is too short now to put it in a
single file. I move it into `schema.js` (whose folder is from `graphql/query/`
into `graphql/`, and this change make me also change file `main.js`'s import
statement).

I am not sure why there need be a `readme.md` in `graphql`. So I remove it (It
will be a better choice that put it into `doc` folder of the project root).
