--- back/graphql/readme.md
@@ -1,9 +1,0 @@
-此目录与其子目录用来存放支持项目中GraphQL运行的基础组件。
-
-### 各JS文件的用途
-
-|      文件       |         描述         |
-|:-------------:|:------------------:|
-| `dataType.js` | 用于存放定义的GraphQL对象类型 |
-|  `query.js`   | 用于存放各种GraphQL查询操作  |
-|  `schema.js`  | 用来存放不同种类的GraphQL查询 |
